feat(refactor): AS-817 data category init() → catalog migration
Migrates the data category's per-type wiring from `init()` bodies in
internal/aws/<svc>.go (and matching `*_related.go`) into the catalog
struct literals in internal/aws/catalog_{data,databases,compute}.go.
Per-category PR after AS-807 (compute), AS-808 (containers), AS-810
(databases), AS-811 (monitoring), AS-812 (messaging), AS-813 (secrets).
Spec: docs/refactor/AS-795-init-cycle-break.md §3 + §5.2.

Catalog entries populated for the 5 top-level data resource types named
in the AS-795l scope (s3, athena, glue, ebs, ebs-snap):

- glue (catalog_data.go): Fetcher, FieldKeys, Related (w/ explicit
  ct-events RelatedDef via ctEventsCheckerFor), Navigable, Wave2
  (EnrichGlueJobStatus, prio 10), IssueEnricherFieldKeys (["last_run"]).
- athena (catalog_data.go): Fetcher, FieldKeys, Related (w/ ct-events),
  Wave2 (EnrichAthenaWorkGroup, prio 100). No Navigable in the legacy
  init() body — preserved.
- s3 (catalog_databases.go, identity-only since AS-810): Fetcher
  (uses FetchS3BucketsPageWithNotifications so the lambda/sns/sqs
  notification pivots resolve — verbatim from s3.go init), FieldKeys,
  Related (w/ ct-events), Wave2 (EnrichS3PublicAccessBlock, prio 100),
  IssueEnricherFieldKeys (["status"]). AS-810 explicitly deferred s3
  to "a later sibling" — this is that sibling.
- ebs (catalog_compute.go, identity-only since AS-807): Fetcher,
  FieldKeys, Related (w/ ct-events), Navigable, Wave2
  (EnrichEBSVolumeStatus, prio 10). AS-807 explicitly deferred ebs +
  ebs-snap to "a later PR" — this is that PR.
- ebs-snap (catalog_compute.go, identity-only since AS-807): Fetcher,
  FieldKeys, FetchByIDs (related-panel lazy-add path bypasses the
  OwnerIds=self filter), Related (w/ ct-events), Navigable. Wave2 stays
  nil — only NoOpIssueEnricher is registered (matches AS-810
  redis/redshift, AS-813 secrets/ssm precedent: spec §5.2 template).

Files with `func init()` removed:
- internal/aws/{athena,glue,ebs,s3}.go — Fetcher / FieldKeys / Related /
  Navigable / FetchByIDs now in catalog.
- internal/aws/s3_related.go — `RegisterRelated("s3", …)` moved to
  catalog_databases.go s3 entry.

s3_objects child type extracted to its own file (sibling precedent
mirrors asg_activities.go / ecs_svc_logs.go / lambda_invocations.go):
- New file internal/aws/s3_objects.go owns the s3_objects child
  registration (RegisterFieldKeys + RegisterPaginatedChild +
  RegisterChildType). This lets s3.go drop its init() body per the
  AS-817 acceptance grep while keeping s3_objects functional until
  AS-795n migrates child types into catalog.SetChildTypes.

Files with init() body retained for legacy consumers (same rationale as
AS-807/AS-810/AS-811/AS-812/AS-813):
- internal/aws/{athena,glue,ebs,s3}_issue_enrichment.go — real Enrich*
  registered into the legacy IssueEnricherRegistry. Still iterated by
  BuildEnrichQueue / runtime_adapter_navigate / probe_adapter /
  TestAttentionSignalsDoc; AS-795n migrates those consumers and deletes
  the legacy registry. glue_issue_enrichment.go and s3_issue_enrichment.go
  also still own RegisterIssueEnricherFieldKeys (install.go bridge does
  not yet wire that catalog field — also AS-795n scope).
- internal/aws/ebs_snap_issue_enrichment.go — registers
  NoOpIssueEnricher so TestConformance_EveryResourceTypeHasWave2Registration
  sees an explicit entry for ebs-snap; Wave2 stays nil per template.

Out of scope per sibling precedent (AS-812 eb_rule_targets /
AS-813 acm-ses interpretation):
- internal/aws/glue_runs.go (child type) — its init() body stays.
  Child types still register via internal/resource; install.go has no
  bridge for ChildFetcher yet. AS-795n (Wave 2 / child infrastructure
  cleanup) migrates ecs_svc_logs / ecs_svc_tasks / lambda_invocations /
  asg_activities / glue_runs / s3_objects / cb_builds / etc. in one
  sweep with the matching install.go and consumer updates. The AS-817
  issue body's literal acceptance grep lists glue_runs.go but the
  AS-812 PR #402 description called out the same shape for
  eb_rule_targets.go as out-of-scope for the per-category PR.
- internal/aws/ebs_snap.go — file does not exist; the ebs and ebs-snap
  registrations both lived in ebs.go (which had init() removed). Spec
  drafting artifact in the AS-817 acceptance grep.

Acceptance per AS-817 issue body:
- rg '^func init\(\)' internal/aws/{s3,athena,glue,ebs}.go and
  internal/aws/s3_related.go → zero hits.
- rg 'Fetcher:' internal/aws/catalog_data.go → 2 hits (glue, athena).
- rg 'Fetcher:' internal/aws/catalog_databases.go → 9 hits (was 8 in
  AS-810; +1 = s3).
- rg 'Fetcher:' internal/aws/catalog_compute.go → 8 hits (was 6 in
  AS-807; +2 = ebs, ebs-snap).
- go build, go test ./tests/unit/, go test -tags integration
  ./tests/integration/, golangci-lint, govulncheck, gofix,
  verify-readonly, Golden|Scenario snapshot, Visual|Scenario
  integration snapshot all pass locally. test-race blocked by
  missing cgo/gcc in this pod (documented under AS-149); mdlint
  N/A — no .md files modified by this PR.

Parent: AS-805 (umbrella).
Depends on: AS-795a / PR #392 (merged), AS-807 / PR #393 (merged),
AS-810 / PR #396 (merged).
