+++ internal/aws/s3.go
@@ -3,91 +3,12 @@ package aws
 import (
 	"context"
 	"fmt"
-	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 
 	"github.com/k2m30/a9s/v3/internal/resource"
 )
-
-func init() {
-	resource.RegisterFieldKeys("s3", []string{
-		"name",
-		"bucket_name",
-		"creation_date",
-		"notification_lambda",
-		"notification_sqs",
-		"notification_sns",
-	})
-
-	resource.RegisterFieldKeys("s3_objects", []string{
-		"key",
-		"size",
-		"last_modified",
-		"storage_class",
-	})
-
-	resource.RegisterPaginated("s3", func(ctx context.Context, clients any, continuationToken string) (resource.FetchResult, error) {
-		c, ok := clients.(*ServiceClients)
-		if !ok || c == nil {
-			return resource.FetchResult{}, fmt.Errorf("AWS clients not initialized")
-		}
-		// Related-panel contract (docs/resources/s3.md §2): lambda/sns/sqs
-		// pivots must resolve non-zero when this bucket has a matching
-		// notification target. Those checkers read Fields["notification_*"],
-		// which can only be populated by GetBucketNotificationConfiguration
-		// — so the list path must run it per-bucket. Accepts N+1 per page
-		// (cheap API, typically ≤50 buckets per AWS account) in exchange
-		// for having the notification pivots actually work.
-		return FetchS3BucketsPageWithNotifications(ctx, c.S3, c.S3, continuationToken)
-	})
-
-	// Register S3 objects as a child type with its own fetcher.
-	resource.RegisterPaginatedChild("s3_objects", func(ctx context.Context, clients any, parentCtx resource.ParentContext, continuationToken string) (resource.FetchResult, error) {
-		c, ok := clients.(*ServiceClients)
-		if !ok || c == nil {
-			return resource.FetchResult{}, fmt.Errorf("AWS clients not initialized")
-		}
-		return FetchS3Objects(ctx, c.S3, parentCtx["bucket"], parentCtx["prefix"], continuationToken)
-	})
-	resource.RegisterChildType(resource.ResourceTypeDef{
-		Name:      "S3 Objects",
-		ShortName: "s3_objects",
-		Columns:   resource.S3ObjectColumns(),
-		Children: []resource.ChildViewDef{{
-			ChildType:      "s3_objects",
-			Key:            "enter",
-			ContextKeys:    map[string]string{"bucket": "@parent.bucket", "prefix": "ID"},
-			DisplayNameKey: "bucket",
-			DrillCondition: func(r resource.Resource) bool { return r.Status == "folder" },
-		}},
-		// RelatedContextFromIDs extracts the bucket name from related IDs encoded as
-		// "bucket|key". Used when navigating to s3_objects from the related panel
-		// (e.g., from a CloudTrail event detail view).
-		RelatedContextFromIDs: func(relatedIDs []string) map[string]string {
-			for _, id := range relatedIDs {
-				parts := strings.SplitN(id, "|", 2)
-				if len(parts) != 2 || parts[0] == "" {
-					continue
-				}
-				bucket := parts[0]
-				key := parts[1]
-				// Derive the prefix (folder path) from the key so the child view
-				// lands on the folder containing the object, not the bucket root.
-				// Example: key="prod/config.json" → prefix="prod/"
-				// Example: key="landing/2026/04/07/x.parquet" → prefix="landing/2026/04/07/"
-				// Example: key="build-4821.tar.gz" → prefix=""
-				prefix := ""
-				if idx := strings.LastIndex(key, "/"); idx >= 0 {
-					prefix = key[:idx+1]
-				}
-				return map[string]string{"bucket": bucket, "prefix": prefix}
-			}
-			return map[string]string{"bucket": "", "prefix": ""}
-		},
-	})
-}
 
 // FetchS3Buckets calls the S3 ListBuckets API and returns all pages of buckets.
 // Used by tests; the production path uses the per-page fetcher for pagination.
